requirements.txt
Django
python-telegram-bot
channels
channels_redis